Define egor glaobal step
